// app/constants/enums/http-codes.ts
export enum HTTPCode {
    OK = 200,
    Created = 201,
    PermanentRedirect = 301,
    TemporaryRedirect = 302,
    BadRequest = 400,
    Unauthorised = 401,
    PaymentRequired = 402,
    Forbidden = 403,
    NotFound = 404,
    MethodNotAllowed = 405,
    Gone = 410,
    PreconditionFailed = 412,
    PayloadTooLarge = 413,
    UriTooLong = 414,
    UnsupportedMediaType = 415,
    InternalServerError = 500,
    NotImplemented = 501,
    BadGateway = 502,
    ServiceUnavailable = 503,
    GatewayTimeout = 504,
    HTTPVersionNotSupported = 505
}